Organize FRAPI mixins
