+++ fabric-renderer-api-v1/src/client/java/net/fabricmc/fabric/mixin/renderer/client/render/MooshroomMushroomFeatureRendererMixin.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.mixin.renderer.client;
+package net.fabricmc.fabric.mixin.renderer.client.render;
 
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
